% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GimmmeMySC-package.R
\docType{package}
\name{GimmmeMySC-package}
\alias{GimmmeMySC}
\alias{GimmmeMySC-package}
\title{GimmmeMySC: Single-Cell Data Utilities}
\description{
End-to-end workflow for the analysis of single-cell and single-nucleus omics data, including scRNA-seq and scATAC-seq. Features include rigorous quality control, ambient RNA correction, and advanced detection of doublets and duplicated cells. Offers tools for normalization, feature selection, dimensionality reduction, clustering, and robust batch effect correction. Supports cell-type annotation using reference atlases and marker-based methods, sex-specific marker analysis, and cell cycle scoring. Enables multi-omic integration, pseudo-bulk analysis, differential expression and accessibility testing, peak calling, motif enrichment, and gene regulatory network reconstruction. Includes extensive visualization capabilities such as UMAP/t-SNE embeddings, violin and dot plots, coverage maps, co-accessibility networks, and interactive heatmaps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecamenen/GimmmeMySC}
  \item Report bugs at \url{https://github.com/ecamenen/GimmmeMySC/issues}
}

}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com} (\href{https://orcid.org/0000-0002-2398-0190}{ORCID})

}
\keyword{internal}
